requirements.txt
playwright
robotframework
robotframework-seleniumlibrary
slack-sdk
httpx
langchain
openai  # optional if using OpenAI instead of Gemini
python-dotenv